Move Vuex plugins into separate file
